add ExceptionHandler and incView

+++ src/main/java/com/y/community/dto/QuestionDTO.java
@@ -12,7 +12,7 @@ public class QuestionDTO extends Question {
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Long creator;
+    private Integer creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
